Move ToolContext -> io_context.py
